+++ src/components/SideNavigationMenu/SideNavigationMenu.tsx
@@ -3,7 +3,7 @@ import CalendarCheckIcon from "@/assets/svgs/ic_calendar_check.svg";
 import SettingCheckIcon from "@/assets/svgs/ic_setting_check.svg";
 import TextboxCheckIcon from "@/assets/svgs/ic_textbox_check.svg";
 import ProfileUpload from "@/src/components/SideNavigationMenu/Common/ProfileUpload";
-import useUserStore from "@/src/stores/userStore";
+import useUserStore from "@/src/stores/useUserStore";
 import clsx from "clsx";
 import Image from "next/image";
 import Link from "next/link";
