+++ src/pages/activity/[[...activityId]]/index.tsx
@@ -8,7 +8,7 @@ import MyActivityHandler from "@/src/components/MyActivities/MyActivityHandler";
 import Review from "@/src/components/Review";
 import Custom404 from "@/src/pages/404";
 import { useGetActivities } from "@/src/queries/useActivities";
-import useUserStore from "@/src/stores/userStore";
+import useUserStore from "@/src/stores/useUserStore";
 import { ActivityDetailResponse } from "@/src/types/activitiesResponses";
 import { useRouter } from "next/router";
