+++ src/pages/my-activities/registration/index.tsx
@@ -1,6 +1,6 @@
 import ActivityConservation from "@/src/components/MyActivities/ActivityConservation";
 import SideNavigationMenu from "@/src/components/SideNavigationMenu/SideNavigationMenu";
-import useUserStore from "@/src/stores/userStore";
+import useUserStore from "@/src/stores/useUserStore";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
