+++ src/pages/sign-in/index.tsx
@@ -7,7 +7,7 @@ import PATH_NAMES from "@/src/constants/pathname";
 import TEST_USERS from "@/src/constants/testUsers";
 import SignPageWrap from "@/src/containers/SignPageWrap";
 import { useSignIn } from "@/src/queries/auth";
-import useUserStore from "@/src/stores/userStore";
+import useUserStore from "@/src/stores/useUserStore";
 import { SignInProps } from "@/src/types/user";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
