+++ src/pages/sign-up/index.tsx
@@ -6,7 +6,7 @@ import PATH_NAMES from "@/src/constants/pathname";
 import SignPageWrap from "@/src/containers/SignPageWrap";
 import { useSignUp } from "@/src/queries/auth";
 import useTempEmailStore from "@/src/stores/useTempEmailStore";
-import useUserStore from "@/src/stores/userStore";
+import useUserStore from "@/src/stores/useUserStore";
 import { SignUpProps } from "@/src/types/user";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
